+++ external/docker-integration-tests/src/test/scala/org/apache/spark/sql/jdbc/DockerKrbJDBCIntegrationSuite.scala
@@ -26,11 +26,12 @@ import scala.io.Source
 
 import org.apache.hadoop.minikdc.MiniKdc
 
+import org.apache.spark.MiniKDCHelper
 import org.apache.spark.sql.Row
 import org.apache.spark.sql.types.StringType
 import org.apache.spark.util.{SecurityUtils, Utils}
 
-abstract class DockerKrbJDBCIntegrationSuite extends DockerJDBCIntegrationSuite {
+abstract class DockerKrbJDBCIntegrationSuite extends DockerJDBCIntegrationSuite with MiniKDCHelper {
   private var kdc: MiniKdc = _
   protected var entryPointDir: File = _
   protected var initDbDir: File = _
@@ -42,15 +43,8 @@ abstract class DockerKrbJDBCIntegrationSuite extends DockerJDBCIntegrationSuite 
 
   override def beforeAll(): Unit = {
     SecurityUtils.setGlobalKrbDebug(true)
-
-    val kdcDir = Utils.createTempDir()
-    val kdcConf = MiniKdc.createConf()
-    kdcConf.setProperty(MiniKdc.DEBUG, "true")
-    kdc = new MiniKdc(kdcConf, kdcDir)
-    kdc.start()
-
+    kdc = startMiniKdc()
     principal = s"$userName@${kdc.getRealm}"
-
     entryPointDir = Utils.createTempDir()
     initDbDir = Utils.createTempDir()
     val keytabFile = new File(initDbDir, keytabFileName)
